+++ src/Parser/DemonHunter/Vengeance/Modules/PainChart/Pain.js
@@ -4,7 +4,7 @@ import React from 'react';
 import PropTypes from 'prop-types';
 import {Line} from 'react-chartjs-2';
 
-import fetchWcl from 'common/fetchWcl';
+import fetchWcl from 'common/fetchWclApi';
 import {formatDuration} from 'common/format';
 import SPELLS from 'common/SPELLS';
 import ManaStyles from 'Interface/Others/ManaStyles';
